added function to get next rise and set times. tidied up UTC calcs
Due to home.date changing each time a next_rising/setting is calculated,
the times where not being calculated correctly.  Added a separate
function to return those times, ensuring that home.date was reset
between each calc and before returning.

Because the function is called before anything else, the need to test
for UTC time is redundant, so removed that part.  Instead, test for ~UTC
and only then call the local time function.